Moved the FederatedPeg project into the FederationGateway project
