+++ src/Stratis.FederatedPeg.Features.FederationGateway/CounterChain/CounterChainSessionManager.cs
@@ -14,6 +14,7 @@ using Stratis.Bitcoin.Interfaces;
 using Stratis.Bitcoin.P2P.Peer;
 using Stratis.Bitcoin.Utilities;
 using Stratis.FederatedPeg.Features.FederationGateway.Interfaces;
+using Stratis.FederatedPeg.Features.FederationGateway.NetworkHelpers;
 using Recipient = Stratis.FederatedPeg.Features.FederationGateway.Wallet;
 using TransactionBuildContext = Stratis.FederatedPeg.Features.FederationGateway.Wallet.TransactionBuildContext;
