+++ src/Stratis.FederatedPeg.Features.FederationGateway/MonitorChain/Auditors/JsonCrossChainTransactionAuditor.cs
@@ -2,6 +2,7 @@
 using NBitcoin;
 using Stratis.Bitcoin.Configuration;
 using Stratis.Bitcoin.Utilities;
+using Stratis.FederatedPeg.Features.FederationGateway.NetworkHelpers;
 
 namespace Stratis.FederatedPeg.Features.FederationGateway
 {
