+++ src/Stratis.FederatedPeg.Features.FederationGateway/MonitorChain/CrossChainTransactionMonitor.cs
@@ -7,6 +7,7 @@ using Newtonsoft.Json;
 using Stratis.Bitcoin.Interfaces;
 using Stratis.Bitcoin.Utilities;
 using Stratis.FederatedPeg.Features.FederationGateway.MonitorChain;
+using Stratis.FederatedPeg.Features.FederationGateway.NetworkHelpers;
 
 namespace Stratis.FederatedPeg.Features.FederationGateway
 {
