+++ src/Stratis.FederatedPeg.Features.FederationGateway/MonitorChain/MonitorChainSessionManager.cs
@@ -11,6 +11,7 @@ using NBitcoin.JsonConverters;
 using Newtonsoft.Json;
 using Stratis.Bitcoin.Interfaces;
 using Stratis.FederatedPeg.Features.FederationGateway.Models;
+using Stratis.FederatedPeg.Features.FederationGateway.NetworkHelpers;
 
 namespace Stratis.FederatedPeg.Features.FederationGateway.MonitorChain
 {
