+++ src/Stratis.FederatedPeg.Features.FederationGateway/MonitorChain/OpReturnDataReader.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Text;
 using Microsoft.Extensions.Logging;
 using NBitcoin;
+using Stratis.FederatedPeg.Features.FederationGateway.NetworkHelpers;
 
 namespace Stratis.FederatedPeg.Features.FederationGateway
 {
