+++ src/Stratis.FederatedPeg.Features.FederationGateway/NetworkHelpers/Chain.cs
@@ -1,0 +1,11 @@
+﻿namespace Stratis.FederatedPeg.Features.FederationGateway.NetworkHelpers
+{
+    /// <summary>
+    /// The two chains that participate in an cross chain transaction.
+    /// </summary>
+    public enum Chain
+    {
+        Mainchain,
+        Sidechain
+    }
+}
