--- src/Stratis.FederatedPeg.Features.SidechainGeneratorServices/Stratis.FederatedPeg.Features.SidechainGeneratorServices.csproj
@@ -1,23 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup Label="Globals">
-    <SccProjectName>SAK</SccProjectName>
-    <SccProvider>SAK</SccProvider>
-    <SccAuxPath>SAK</SccAuxPath>
-    <SccLocalPath>SAK</SccLocalPath>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Stratis.Bitcoin" Version="1.1.11-beta" />
-    <PackageReference Include="Stratis.Bitcoin.Features.Miner" Version="1.1.11-beta" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Stratis.FederatedPeg\Stratis.FederatedPeg.csproj" />
-  </ItemGroup>
-
-</Project>
