--- src/Stratis.FederatedPeg/AssemblyInfo.cs
@@ -1,4 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("Stratis.FederatedPeg.Tests")]
-[assembly: InternalsVisibleTo("Stratis.FederatedPeg.IntegrationTests")]
