--- src/Stratis.FederatedPeg/NetworkHelpers/Chain.cs
@@ -1,11 +1,0 @@
-﻿namespace Stratis.FederatedPeg
-{
-    /// <summary>
-    /// The two chains that participate in an cross chain transaction.
-    /// </summary>
-    public enum Chain
-    {
-        Mainchain,
-        Sidechain
-    }
-}
