--- src/Stratis.FederatedPeg/Stratis.FederatedPeg.csproj
@@ -1,27 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup Label="Globals">
-    <SccProjectName>SAK</SccProjectName>
-    <SccProvider>SAK</SccProvider>
-    <SccAuxPath>SAK</SccAuxPath>
-    <SccLocalPath>SAK</SccLocalPath>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="NStratis" Version="4.0.0.60" />
-    <PackageReference Include="Stratis.Bitcoin" Version="1.1.11-beta" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Stratis.Sidechains.Networks\Stratis.Sidechains.Networks.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="MemberFolderManagers\" />
-  </ItemGroup>
-
-</Project>
