+++ src/Stratis.FederatedSidechains.IntegrationTests/GatewayIntegrationTestEnvironment.cs
@@ -15,6 +15,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Stratis.Bitcoin.IntegrationTests.Common;
+using Stratis.FederatedPeg.Features.FederationGateway.NetworkHelpers;
 
 namespace Stratis.FederatedSidechains.IntegrationTests
 {
